# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Black Tie VoIP Reseller Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reseller/check-domain:
    get:
      operationId: resellerCheckDomain
      tags: [reseller]
      summary: Check domain availability via RDAP
      parameters:
        - name: domain
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Domain availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainCheckResult"

  # ── DID Manager ──────────────────────────────────────────────────────────────
...
  /reseller/clients:
    get:
      operationId: resellerGetClients
      tags: [resellers]
      summary: Get reseller's clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: resellerCreateClient
      tags: [resellers]
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /reseller/clients/{id}:
    get:
      operationId: resellerGetClient
      tags: [resellers]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    put:
      operationId: resellerUpdateClient
      tags: [resellers]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientRequest"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: resellerDeleteClient
      tags: [resellers]
      summary: Delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /reseller/profile:
    get:
      operationId: resellerGetProfile
      tags: [resellers]
      summary: Get reseller profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reseller"
    put:
      operationId: resellerUpdateProfile
      tags: [resellers]
      summary: Update reseller profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResellerProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reseller"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Reseller:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        unitStreetNumber:
          type: string
        buildingComplex:
          type: string
        streetName:
          type: string
        address:
          type: string
        address2:
          type: string
        city:
          type: string
        province:
          type: string
        commissionRate:
          type: number
        status:
          type: string
          enum: [active, suspended, pending]
        totalClients:
          type: integer
        monthlyRevenue:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, companyName, contactName, email, status, commissionRate, totalClients, monthlyRevenue, createdAt]
...
    UpdateResellerProfileRequest:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        unitStreetNumber:
          type: string
        buildingComplex:
          type: string
        streetName:
          type: string
        address:
          type: string
        address2:
          type: string
        city:
          type: string
        province:
          type: string
    Client:
      type: object
      properties:
        id:
          type: integer
        resellerId:
          type: integer
        resellerName:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        unitStreetNumber:
          type: string
        buildingComplex:
          type: string
        streetName:
          type: string
        address:
          type: string
        address2:
          type: string
        city:
          type: string
        province:
          type: string
        sipExtensions:
          type: integer
        monthlyFee:
          type: number
        status:
          type: string
          enum: [active, suspended, cancelled]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, resellerId, companyName, contactName, email, status, sipExtensions, monthlyFee, createdAt]
    CreateClientRequest:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        unitStreetNumber:
          type: string
        buildingComplex:
          type: string
        streetName:
          type: string
        address:
          type: string
        address2:
          type: string
        city:
          type: string
        province:
          type: string
        sipExtensions:
          type: integer
        monthlyFee:
          type: number
        notes:
          type: string
      required: [companyName, contactName, email, sipExtensions, monthlyFee]
    UpdateClientRequest:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        unitStreetNumber:
          type: string
        buildingComplex:
          type: string
        streetName:
          type: string
        address:
          type: string
        address2:
          type: string
        city:
          type: string
        province:
          type: string
        sipExtensions:
          type: integer
        monthlyFee:
          type: number
        status:
          type: string
          enum: [active, suspended, cancelled]
        notes:
          type: string
...
    DomainCheckResult:
      type: object
      properties:
        domain:
          type: string
        available:
          type: boolean
        status:
          type: string
        registrar:
          type: string
        expiresAt:
          type: string
        registrationStatus:
          type: string
      required: [domain, available, status]