# lib/api-zod/src/generated/types/updateResellerProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Black Tie VoIP Reseller Portal API
 * OpenAPI spec version: 0.2.0
 */

export interface UpdateResellerProfileRequest {
  companyName?: string;
  contactName?: string;
  email?: string;
  phone?: string;
  unitStreetNumber?: string;
  buildingComplex?: string;
  streetName?: string;
  address?: string;
  address2?: string;
  city?: string;
  province?: string;
}